requirements.txt
python-telegram-bot==20.7
google-generativeai==0.7.2
tqdm==4.66.5
tenacity==8.5.0
requests==2.32.3
